# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AugmentedQuill",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/projects/delete": {
      "post": {
        "tags": ["Projects"],
        "summary": "Api Projects Delete",
        "description": "Handle the API request to projects delete.",
        "operationId": "api_projects_delete_api_v1_projects_delete_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectDeleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectMutationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/select": {
      "post": {
        "tags": ["Projects"],
        "summary": "Api Projects Select",
        "description": "Handle the API request to projects select.",
        "operationId": "api_projects_select_api_v1_projects_select_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectSelectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectSelectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/create": {
      "post": {
        "tags": ["Projects"],
        "summary": "Api Projects Create",
        "description": "Handle the API request to projects create.",
        "operationId": "api_projects_create_api_v1_projects_create_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectMutationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/convert": {
      "post": {
        "tags": ["Projects"],
        "summary": "Api Projects Convert",
        "description": "Handle the API request to projects convert.",
        "operationId": "api_projects_convert_api_v1_projects_convert_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectConvertRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectMutationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/books/create": {
      "post": {
        "tags": ["Projects"],
        "summary": "Api Books Create",
        "description": "Handle the API request to books create.",
        "operationId": "api_books_create_api_v1_books_create_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookMutationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/books/delete": {
      "post": {
        "tags": ["Projects"],
        "summary": "Api Books Delete",
        "description": "Handle the API request to books delete.",
        "operationId": "api_books_delete_api_v1_books_delete_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookDeleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookMutationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/books/restore": {
      "post": {
        "tags": ["Projects"],
        "summary": "Api Books Restore",
        "description": "Handle the API request to books restore.",
        "operationId": "api_books_restore_api_v1_books_restore_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookRestoreRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookMutationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/projects/images/update_description": {
      "post": {
        "tags": ["Projects"],
        "summary": "Api Projects Images Update Description",
        "description": "Handle the API request to projects images update description.",
        "operationId": "api_projects_images_update_description_api_v1_projects_images_update_description_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImageDescriptionUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/images/create_placeholder": {
      "post": {
        "tags": ["Projects"],
        "summary": "Api Projects Images Create Placeholder",
        "description": "Handle the API request to projects images create placeholder.",
        "operationId": "api_projects_images_create_placeholder_api_v1_projects_images_create_placeholder_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImagePlaceholderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/images/upload": {
      "post": {
        "tags": ["Projects"],
        "summary": "Api Projects Images Upload",
        "description": "Handle the API request to projects images upload.",
        "operationId": "api_projects_images_upload_api_v1_projects_images_upload_post",
        "parameters": [
          {
            "name": "target_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Target Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_api_projects_images_upload_api_v1_projects_images_upload_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/images/delete": {
      "post": {
        "tags": ["Projects"],
        "summary": "Api Projects Images Delete",
        "description": "Handle the API request to projects images delete.",
        "operationId": "api_projects_images_delete_api_v1_projects_images_delete_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImageDeleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/images/restore": {
      "post": {
        "tags": ["Projects"],
        "summary": "Api Projects Images Restore",
        "description": "Handle the API request to projects images restore.",
        "operationId": "api_projects_images_restore_api_v1_projects_images_restore_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImageRestoreRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/images/{filename}": {
      "get": {
        "tags": ["Projects"],
        "summary": "Api Projects Images Get",
        "description": "Handle the API request to projects images get.",
        "operationId": "api_projects_images_get_api_v1_projects_images__filename__get",
        "parameters": [
          {
            "name": "filename",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Filename"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Api Projects Images Get Api V1 Projects Images  Filename  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/export": {
      "get": {
        "tags": ["Projects"],
        "summary": "Api Projects Export",
        "description": "Handle the API request to projects export.",
        "operationId": "api_projects_export_api_v1_projects_export_get",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Api Projects Export Api V1 Projects Export Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/export/epub": {
      "get": {
        "tags": ["Projects"],
        "summary": "Api Projects Export Epub",
        "description": "Handle the API request to projects export epub.",
        "operationId": "api_projects_export_epub_api_v1_projects_export_epub_get",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Api Projects Export Epub Api V1 Projects Export Epub Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/import": {
      "post": {
        "tags": ["Projects"],
        "summary": "Api Projects Import",
        "description": "Handle the API request to projects import.",
        "operationId": "api_projects_import_api_v1_projects_import_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_api_projects_import_api_v1_projects_import_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectMutationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body_api_projects_images_upload_api_v1_projects_images_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": ["file"],
        "title": "Body_api_projects_images_upload_api_v1_projects_images_upload_post"
      },
...
      "BookCreateRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": ["name"],
        "title": "BookCreateRequest",
        "description": "Represents the BookCreateRequest type."
      },
      "BookDeleteRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": ["name"],
        "title": "BookDeleteRequest",
        "description": "Represents the BookDeleteRequest type."
      },
      "BookMutationResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "book_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Book Id"
          },
          "restore_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Restore Id"
          },
          "story": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StoryPayload"
              },
              {
                "type": "null"
              }
            ]
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          }
        },
        "type": "object",
        "required": ["ok"],
        "title": "BookMutationResponse",
        "description": "Response body for book create/delete/restore endpoints."
      },
      "BookRestoreRequest": {
        "properties": {
          "restore_id": {
            "type": "string",
            "title": "Restore Id"
          }
        },
        "type": "object",
        "required": ["restore_id"],
        "title": "BookRestoreRequest",
        "description": "Represents the BookRestoreRequest type."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ImageDeleteRequest": {
        "properties": {
          "filename": {
            "type": "string",
            "title": "Filename"
          }
        },
        "type": "object",
        "required": ["filename"],
        "title": "ImageDeleteRequest",
        "description": "Represents the ImageDeleteRequest type."
      },
      "ImageDescriptionUpdateRequest": {
        "properties": {
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "default": ""
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title",
            "default": ""
          }
        },
        "type": "object",
        "required": ["filename"],
        "title": "ImageDescriptionUpdateRequest",
        "description": "Represents the ImageDescriptionUpdateRequest type."
      },
      "ImagePlaceholderRequest": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name",
            "default": ""
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title",
            "default": ""
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "default": ""
          }
        },
        "type": "object",
        "title": "ImagePlaceholderRequest",
        "description": "Represents the ImagePlaceholderRequest type."
      },
      "ImageRestoreRequest": {
        "properties": {
          "restore_id": {
            "type": "string",
            "title": "Restore Id"
          }
        },
        "type": "object",
        "required": ["restore_id"],
        "title": "ImageRestoreRequest",
        "description": "Represents the ImageRestoreRequest type."
      },
...
      "ProjectConvertRequest": {
        "properties": {
          "target_type": {
            "type": "string",
            "title": "Target Type"
          }
        },
        "type": "object",
        "required": ["target_type"],
        "title": "ProjectConvertRequest",
        "description": "Represents the ProjectConvertRequest type."
      },
      "ProjectCreateRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language",
            "default": "en"
          }
        },
        "type": "object",
        "required": ["name", "type"],
        "title": "ProjectCreateRequest",
        "description": "Represents the ProjectCreateRequest type."
      },
      "ProjectDeleteRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": ["name"],
        "title": "ProjectDeleteRequest",
        "description": "Represents the ProjectDeleteRequest type."
      },
...
      "ProjectMutationResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "registry": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProjectRegistry"
              },
              {
                "type": "null"
              }
            ]
          },
          "story": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StoryPayload"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": ["ok"],
        "title": "ProjectMutationResponse",
        "description": "Generic response for project create/delete/convert operations."
      },
...
      "ProjectSelectRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": ["name"],
        "title": "ProjectSelectRequest",
        "description": "Represents the ProjectSelectRequest type."
      },
      "ProjectSelectResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "registry": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProjectRegistry"
              },
              {
                "type": "null"
              }
            ]
          },
          "story": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StoryPayload"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          }
        },
        "type": "object",
        "required": ["ok"],
        "title": "ProjectSelectResponse",
        "description": "Response body for ``POST /api/v1/projects/select``."
      },
...
    }
  }
}